refactor(csa-client): PR #683 review コメント 4 件を一括対応 (#682)
claude[bot] からの review コメント 4 件 (軽微) を反映:

1. **protocol.rs:625** — `extract_retry_after_sec` の split を `is_whitespace()
   || ',' || ';'` から `split_whitespace()` に簡略化。CSA プロトコルでは
   whitespace 以外のトークン区切りは存在しないため防御的実装は YAGNI。
2. **rate_limit_retry_after.rs:73** — `compute_effective_retry_delay` を
   `protocol.rs` に移動し `pub fn` 化、lib root から re-export。integration
   test 側で再実装していた `effective_delay` を削除し本物を直接 import。仕様
   変更時の二重メンテナンスを解消、`main.rs` 側との乖離リスクを排除。
3. **main.rs:348** — 両 Err arm で重複していた「warn/info 出し分け +
   `sleep_with_shutdown`」ブロックを `sleep_retry()` helper に集約。将来の
   ログフォーマット変更で片方の修正が漏れるリスクを解消。
4. **main.rs:238** — `checked_add` overflow → `deadline = None` 時の挙動
   ('shutdown まで `SHUTDOWN_POLL` 刻みで polling 継続') を inline コメントで
   明記。実害は低いが意図を明確化。

test:
- `protocol::tests` に `compute_effective_retry_delay_*` 4 件を移植
  (extract_retry_after_sec の test と同居)。
- `main::tests` から同 4 件を削除し `sleep_with_shutdown_*` 3 件のみ残す。
- integration test 5 件は `compute_effective_retry_delay` を直接呼ぶように更新。
- 全 60 件 (lib 27 + bin 23 + integration 5 + 旧 5) pass、clippy strict
  (`-D warnings`) 通過。

Refs: #682, #683 (review)
